start phase 3

+++ ADCS/TTRD_19/source/adcs/services_utilities/time.h
@@ -87,7 +87,7 @@ typedef struct {
    @return  error code or Zero if no error.
 
 */
-uint8_t time_init(void);
+uint32_t time_init(void);
 
 /*
    time_update
